backend/requirements.txt
gensim>=4.3.0
pandas>=1.3.0
python-dotenv>=0.19.0
openai>=0.27.0
supabase>=0.3.0